dashboard_app/requirements.txt
dash==2.2.0
dash-bootstrap-components==1.0.3
dash-bootstrap-templates==1.0.5
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.0.3
gunicorn==20.1.0
lxml==4.8.0
matplotlib>=3.5
numpy==1.22.2
pandas==1.4.1
Pillow==9.0.1
plotly==5.6.0
psutil==5.9.0
PyYAML==6.0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
SQLAlchemy==1.4.31
statsmodels==0.13.2
ujson==5.1.0
urllib3==1.26.8
werkzeug==2.0.1
psycopg2-binary==2.9.3